refactor : follow, like, user, board Service interface 화
