+++ src/main/java/com/example/foodthought/controller/BookController.java
@@ -1,18 +1,14 @@
 package com.example.foodthought.controller;
 
 import com.example.foodthought.common.dto.ResponseDto;
-import com.example.foodthought.dto.book.CreateBookRequestDto;
 import com.example.foodthought.dto.book.GetBookResponseDto;
-import com.example.foodthought.dto.book.UpdateBookRequestDto;
 import com.example.foodthought.service.BookService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
 import java.util.List;
 
 @Slf4j
