+++ src/main/java/com/example/foodthought/service/BoardService.java
@@ -5,10 +5,7 @@ import com.example.foodthought.dto.board.CreateBoardRequestDto;
 import com.example.foodthought.dto.board.GetBoardResponseDto;
 import com.example.foodthought.dto.board.UpdateBoardRequestDto;
 import com.example.foodthought.entity.*;
-import com.example.foodthought.repository.BoardRepository;
-import com.example.foodthought.repository.BookRepository;
-import com.example.foodthought.repository.CommentRepository;
-import com.example.foodthought.repository.LikeRepository;
+import com.example.foodthought.repository.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
